fix: fix toast message about share function
NFT 헌혈증에서 공유버튼을 눌렀을 때 이미지 다운로드와 토스트 메시지 출력이 모두 잘 진행되도록 토스트 메시지의 위치, 디자인 등을 수정했습니다.
